refactor: unify analytics and send pages
